+++ skeleton_files/bot_code.py
@@ -1,5 +1,5 @@
 class Pot:
-    value: float
+    value: int
     players: list[str]
 
 
@@ -8,12 +8,12 @@ class GameState:
     index_of_small_blind: int
     players: list[str]
     player_cards: list[str]
-    held_money: list[float]
-    bet_money: list[float]
+    held_money: list[int]
+    bet_money: list[int]
     community_cards: list[str]
     pots: list[Pot]
-    small_blind: float
-    big_blind: float
+    small_blind: int
+    big_blind: int
 
 
 # DO NOT CHANGE ABOVE CODE OR FUNCTION SIGNATURE ELSE YOUR CODE WILL NOT RUN!
